refactor(backend): object-lifecycle split and review follow-ups
- Split object_lifecycle.rs into config/service/worker/deletion submodules
  using the project's parent-file convention, preserving the public API via
  re-exports.
- Single-source the upload-session column projection with a macro and trim
  the four queries to the columns actually decoded.
- Redact object name, timestamp, and coordinates from the update-metadata
  debug log, keeping the id and a has_location flag.
- Surface a parked-backlog warning each maintenance pass for cleanups and
  deletions that exhausted their retry budget, backed by two count queries.
- Document the password-reset outbox plaintext-token trade-off, and add the
  incomplete-multipart lifecycle and storage-origin-isolation notes to the
  deployment guide.
